+++ employeesystem/src/main/java/com/employee/EmployeeController.java
@@ -1,6 +1,5 @@
-package com.api;
+package com.employee;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,28 +7,26 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 public class EmployeeController {
 
-    private final EmployeeRepository employeeRepository;
+    private final EmployeeService employeeService;
 
-    @Autowired
-    public EmployeeController(EmployeeRepository employeeRepository) {
-        this.employeeRepository = employeeRepository;
+    public EmployeeController(EmployeeService employeeService) {
+        this.employeeService = employeeService;
     }
 
     @GetMapping("/api/employees")
     public ResponseEntity<List<Employee>> findEmployees() {
-        List<Employee> employees = employeeRepository.findAll();
+        List<Employee> employees = employeeService.findEmployees();
         return new ResponseEntity<>(employees, HttpStatus.OK);
     }
 
     @GetMapping("/api/employees/{employee-id}")
     public ResponseEntity<Employee> findEmployee(@PathVariable("employee-id") int employeeId) {
-        Optional<Employee> employee = employeeRepository.findById(employeeId);
-        return new ResponseEntity<>(employee.orElseThrow(() -> new RuntimeException("Employee does not exist.")), HttpStatus.OK);
+        Employee employee = employeeService.findEmployee(employeeId);
+        return new ResponseEntity<>(employee, HttpStatus.OK);
     }
 
 }
