+++ employeesystem/src/main/java/com/employee/EmployeeRepository.java
@@ -1,4 +1,4 @@
-package com.api;
+package com.employee;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
