--- employeesystem/src/test/java/com/api/MainTest.java
@@ -1,20 +1,0 @@
-package com.api;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertTrue;
-
-/**
- * Unit test for simple Main.
- */
-public class MainTest
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
